+++ opentracing-shim/test/shim_utils_test.cc
@@ -26,6 +26,7 @@
 #include "opentelemetry/common/attribute_value.h"
 #include "opentelemetry/common/key_value_iterable.h"
 #include "opentelemetry/common/timestamp.h"
+#include "opentelemetry/context/context.h"
 #include "opentelemetry/context/runtime_context.h"
 #include "opentelemetry/nostd/function_ref.h"
 #include "opentelemetry/nostd/shared_ptr.h"
@@ -34,13 +35,13 @@
 #include "opentelemetry/opentracingshim/shim_utils.h"
 #include "opentelemetry/opentracingshim/span_context_shim.h"
 #include "opentelemetry/trace/span_context.h"
+#include "opentelemetry/trace/span_metadata.h"
 #include "opentelemetry/trace/span_startoptions.h"
-
-#include "shim_mocks.h"
 
 namespace trace_api = opentelemetry::trace;
 namespace baggage   = opentelemetry::baggage;
 namespace common    = opentelemetry::common;
+namespace context   = opentelemetry::context;
 namespace nostd     = opentelemetry::nostd;
 namespace shim      = opentelemetry::opentracingshim;
 
@@ -138,8 +139,7 @@ TEST(ShimUtilsTest, MakeOptionsShim_EmptyRefs)
             common::SystemTimestamp{options.start_system_timestamp});
   ASSERT_EQ(options_shim.start_steady_time,
             common::SteadyTimestamp{options.start_steady_timestamp});
-  ASSERT_EQ(nostd::get<trace_api::SpanContext>(options_shim.parent),
-            trace_api::SpanContext::GetInvalid());
+  ASSERT_FALSE(nostd::get<context::Context>(options_shim.parent).HasKey(trace_api::kSpanKey));
 }
 
 TEST(ShimUtilsTest, MakeOptionsShim_InvalidSpanContext)
@@ -154,8 +154,8 @@ TEST(ShimUtilsTest, MakeOptionsShim_InvalidSpanContext)
             common::SystemTimestamp{options.start_system_timestamp});
   ASSERT_EQ(options_shim.start_steady_time,
             common::SteadyTimestamp{options.start_steady_timestamp});
-  ASSERT_EQ(nostd::get<trace_api::SpanContext>(options_shim.parent),
-            trace_api::SpanContext::GetInvalid());
+
+  ASSERT_FALSE(nostd::get<context::Context>(options_shim.parent).HasKey(trace_api::kSpanKey));
 }
 
 TEST(ShimUtilsTest, MakeOptionsShim_FirstChildOf)